calculate

+++ Calculator/ViewController.h
@@ -12,9 +12,9 @@
 
 @property NSUInteger isInitialState;//初期状態かどうか　YES：初期　NO：違う
 
-@property NSInteger transrateValue;//文字列を数値に変換したものを格納する変数
+@property CGFloat transrateValue;//文字列を数値に変換したものを格納する変数
 
-@property NSInteger currentValue;//演算子が押されたときに現在の数値を格納する。
+@property CGFloat currentValue;//演算子が押されたときに現在の数値を格納する。
 
 @property NSInteger operation;//どの演算子が押されたかの判別
